Changed a lot of small things across the board, edits to more file than not.
This fixes a few bugs, mainly to due with time breaking the model, and, implements the observer observable framework aspect for the relationship between the model and the view.

Includes an update method for the view, which should work for the current view as well as any children. Seeing as how it just calls the 'setCenter' method, and adds the requirement that all set center methods call the "resetCenter" before doing anything.