Lots of major and minor Changes Major : fixed integration with the main view and the different calenderViews. Rearranged the view slightly to make this possible, changed constructor parameters to take needed information
Minor : bug fixes with the add event modal, the linking of the sidebar date pick with the rest of the app, etc. See specific change notes for all of these. Accounts for BUG_1 and BUG_2 in trello plus more that never made it into trello

This commit represents a mostly function and stable application